Se realizo los cambios a los archivos para la verificacion del lanzamientos de las excpeciones cuando intentamos ingresar un ID que todavia no es valido como el 505

+++ demo/src/main/java/com/example/demo/controller/MenuController.java
@@ -27,7 +27,7 @@ public class MenuController {
     }
 
     @GetMapping("/menu/detalle/{id}")
-    public String detalle(@PathVariable Integer id, Model model) {
+    public String detalle(@PathVariable Long id, Model model) {
 
         Producto producto = productoService.obtenerPorId(id);
